Removed & updated tests.
