+++ tests/e2e/helpers/jsFilesAreValid.js
@@ -1,21 +1,22 @@
 const fs = require('fs-extra')
-const { Linter } = require('eslint')
+const {Linter} = require('eslint')
 const linter = new Linter()
 const filesAndFolders = require('../config/filesAndFolders')
-const { absolutePathConfig, listFolderContents } = require('./folderFileHelper')
+const {absolutePathConfig, listFolderContents} = require('./folderFileHelper')
 const lintConfig = {
   parserOptions: {
     ecmaVersion: 10,
-    sourceType: 'module'
-  }
+    sourceType: 'module',
+  },
 }
 
 function jsFilesAreValid(appPath, appType) {
   const folders = absolutePathConfig(appPath, filesAndFolders[appType])
 
   const jsFiles = Object.keys(folders).reduce((acc, folder) => {
-    const jsFilesInFolder = listFolderContents(folder, { files: true })
-      .filter(file => file.endsWith('.js'))
+    const jsFilesInFolder = listFolderContents(folder, {
+      files: true,
+    }).filter(file => file.endsWith('.js'))
 
     return acc.concat(jsFilesInFolder)
   }, [])
@@ -24,7 +25,7 @@ function jsFilesAreValid(appPath, appType) {
     const fileContents = fs.readFileSync(file, 'utf8')
     const errors = linter.verify(fileContents, lintConfig).filter(e => e.fatal)
 
-    return acc.concat(errors.map(e => ({ ...e, file, /*fileContents*/ })))
+    return acc.concat(errors.map(e => ({...e, file /*fileContents*/})))
   }, [])
 
   return !errors.length || JSON.stringify(errors, null, 2)
